+++ R/remove_attributes.R
@@ -1,6 +1,6 @@
 #' Remove attributes
 #'
-#' This function removes specfied attributes. When applied to a data.frame, it will also
+#' This function removes specified attributes. When applied to a data.frame, it will also
 #' remove recursively the specified attributes to each column of the data.frame.
 #'
 #' @param x an object